+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/textual/TextualReportFacade.kt
@@ -1,25 +1,25 @@
 package io.github.surpsg.deltacoverage.report.textual
 
+import io.github.surpsg.deltacoverage.config.CoverageEntity
 import io.github.surpsg.deltacoverage.lib.kotlin.applyIf
 import io.github.surpsg.deltacoverage.report.ReportBound
 import io.github.surpsg.deltacoverage.report.ReportType
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.BRANCHES_H
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.CLASS_H
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.FAILURE_COV_PATTERN
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.HEADERS
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.INDEXED_HEADERS
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.INSTR_H
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.LINES_H
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.MAX_CLASS_COLUMN_LENGTH
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.NA_VALUE
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.PERCENT_MULTIPLIER
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.SHRINK_PLACEHOLDER
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.SUCCESS_COV_PATTERN
 import io.github.surpsg.deltacoverage.report.textual.TextualReportRenderer.Context
 import java.io.OutputStream
 
 internal object TextualReportFacade {
-
-    private const val MAX_CLASS_COLUMN_LENGTH = 100
-    private const val SHRINK_PLACEHOLDER = "..."
-
-    private const val PERCENT_MULTIPLIER = 100
-
-    private const val NA_VALUE = ""
-
-    private const val SOURCE_H = "Source"
-    private const val CLASS_H = "Class"
-    private const val LINES_H = "Lines"
-    private const val BRANCHES_H = "Branches"
-    private val HEADERS = listOf(SOURCE_H, CLASS_H, LINES_H, BRANCHES_H)
 
     fun generateReport(
         buildContext: BuildContext,
@@ -27,7 +27,7 @@ internal object TextualReportFacade {
         val rawCoverageData: List<RawCoverageData> = buildContext.coverageDataProvider.obtainData()
         val coverageDataValues: List<List<String>> =
             rawCoverageData
-                .sortedByDescending { it.linesRatio }
+                .sortedByDescending { it.lines.ratio }
                 .asSequence()
                 .map { it.toValuesCollection(buildContext) }
                 .toList()
@@ -37,49 +37,86 @@ internal object TextualReportFacade {
                 output = buildContext.outputStream
                 title = buildTitle(buildContext)
                 headers = HEADERS
-                footer = rawCoverageData.computeTotal().toValuesCollection(buildContext)
+                footer = buildFooter(rawCoverageData, buildContext)
                 rows = coverageDataValues
             }
         )
     }
 
+    private fun buildFooter(
+        rawCoverageData: List<RawCoverageData>,
+        buildContext: BuildContext,
+    ): List<List<String>> {
+        val expectedValuesLine = Array(INDEXED_HEADERS.size) { "" }.apply {
+            val classHeaderIndex = INDEXED_HEADERS.getValue(CLASS_H)
+            this[classHeaderIndex] = "Min expected"
+        }
+        sequenceOf(
+            BRANCHES_H to CoverageEntity.BRANCH,
+            LINES_H to CoverageEntity.LINE,
+            INSTR_H to CoverageEntity.INSTRUCTION,
+        )
+            .filter { (_, entity) -> !buildContext.targetCoverage.getOrDefault(entity, Double.NaN).isNaN() }
+            .filter { (_, entity) -> buildContext.targetCoverage.getValue(entity) > 0.0 }
+            .forEach { (headerName, entity) ->
+                val expectedCoverageRatio: Double = buildContext.targetCoverage.getValue(entity)
+                val headerIndex = INDEXED_HEADERS.getValue(headerName)
+                expectedValuesLine[headerIndex] = expectedCoverageRatio.formatToPercentage()
+            }
+
+        return listOf(
+            rawCoverageData.computeTotal().toValuesCollection(buildContext) { coverage, formatted ->
+                val expectedRatio: Double? = buildContext.targetCoverage[coverage.entity]
+                when {
+                    expectedRatio == null -> formatted
+                    coverage.ratio < expectedRatio -> FAILURE_COV_PATTERN.format(formatted)
+                    else -> SUCCESS_COV_PATTERN.format(formatted)
+                }
+            },
+            expectedValuesLine.toList(),
+        )
+    }
+
     private fun RawCoverageData.toValuesCollection(
         buildContext: BuildContext,
-    ): List<String> = HEADERS.map { header ->
+        decorateCoverageValue: (Coverage, String) -> String = { _, formatted -> formatted },
+    ): List<String> = INDEXED_HEADERS.keys.map { header ->
         when (header) {
-            SOURCE_H -> source
-
-            LINES_H -> linesRatio.formatToPercentage()
-
             CLASS_H -> aClass.applyIf(buildContext.shrinkLongClassName) {
                 shrinkClassName(MAX_CLASS_COLUMN_LENGTH)
             }
 
+            LINES_H -> decorateCoverageValue(lines, lines.ratio.formatToPercentage())
+
             BRANCHES_H -> {
-                if (branchesTotal == 0) {
+                if (branches.total == 0) {
                     NA_VALUE
                 } else {
-                    branchesRatio.formatToPercentage()
+                    decorateCoverageValue(branches, branches.ratio.formatToPercentage())
                 }
             }
+
+            INSTR_H -> decorateCoverageValue(instr, instr.ratio.formatToPercentage())
+
 
             else -> error("Unknown header: $header")
         }
     }
 
     private fun List<RawCoverageData>.computeTotal(): RawCoverageData = fold(
-        RawCoverageData.newBlank { group = "Total" },
-        RawCoverageData::merge
+        RawCoverageData.newBlank { aClass = "Total" },
+        RawCoverageData::merge,
     )
 
     private fun Double.formatToPercentage(): String {
-        val percents = this.toFloat() * PERCENT_MULTIPLIER
-        val pattern = if (percents % 1 == 0.0f) {
-            "%.0f"
+        val percents: Double = this * PERCENT_MULTIPLIER
+        val percentsInt = percents.toInt()
+        val pattern = if (percents > percentsInt) {
+            "%.2f%%"
         } else {
-            "%.2f"
+            "%.0f%%"
         }
-        return pattern.format(percents) + '%'
+        return pattern.format(percents)
     }
 
     private fun String.shrinkClassName(maxLength: Int): String {
@@ -103,6 +140,7 @@ internal object TextualReportFacade {
         val coverageDataProvider: RawCoverageDataProvider,
         val outputStream: OutputStream,
         val shrinkLongClassName: Boolean,
+        val targetCoverage: Map<CoverageEntity, Double>,
     ) {
 
         init {
@@ -121,13 +159,22 @@ internal object TextualReportFacade {
             lateinit var outputStream: OutputStream
             var shrinkLongClassName: Boolean = false
 
-            fun build() = BuildContext(
-                reportType,
-                reportBound,
-                coverageDataProvider,
-                outputStream,
-                shrinkLongClassName
-            )
+            private val targetCoverages: MutableMap<CoverageEntity, Double> = mutableMapOf()
+
+            fun targetCoverage(coverageEntity: CoverageEntity, ratio: Double) {
+                targetCoverages[coverageEntity] = ratio
+            }
+
+            fun build(): BuildContext {
+                return BuildContext(
+                    reportType,
+                    reportBound,
+                    coverageDataProvider,
+                    outputStream,
+                    shrinkLongClassName,
+                    targetCoverages,
+                )
+            }
         }
 
         companion object {