+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/textual/asciitable/AsciiTableRenderer.kt
@@ -30,8 +30,10 @@ internal object AsciiTableRenderer : TextualReportRenderer() {
     }
 
     override fun Context.renderFooter(printWriter: PrintWriter, widthMap: List<Int>) = with(printWriter) {
-        printRow(footer, widthMap)
-        printLine(widthMap)
+        multiLineFooter.forEach { footer ->
+            printRow(footer, widthMap)
+            printLine(widthMap)
+        }
     }
 
     @Suppress("ImplicitDefaultLocale")