+++ delta-coverage-core/src/test/kotlin/io/github/surpsg/deltacoverage/report/intellij/coverage/IntellijRawCoverageDataProviderTest.kt
@@ -1,7 +1,10 @@
 package io.github.surpsg.deltacoverage.report.intellij.coverage
 
+import com.intellij.rt.coverage.data.JumpsAndSwitches
 import com.intellij.rt.coverage.data.LineData
 import com.intellij.rt.coverage.data.ProjectData
+import com.intellij.rt.coverage.data.instructions.ClassInstructions
+import com.intellij.rt.coverage.data.instructions.LineInstructions
 import io.github.surpsg.deltacoverage.report.textual.RawCoverageData
 import io.kotest.assertions.assertSoftly
 import io.kotest.matchers.collections.shouldBeEmpty
@@ -29,13 +32,32 @@ class IntellijRawCoverageDataProviderTest {
         // GIVEN
         val className = "com.example.Class1"
         val sourceName = "Class1.java"
+        val instrCovered = 3
+        val instrUncovered = 5
         val provider = IntellijRawCoverageDataProvider(
             ProjectData().apply {
+                instructions[className] = ClassInstructions(
+                    arrayOf(
+                        LineInstructions().apply { instructions = instrCovered },
+                        LineInstructions().apply { instructions = instrUncovered }
+                    )
+                )
+
                 getOrCreateClassData(className).apply {
                     source = sourceName
 
                     setLines(
-                        arrayOf(LineData(1, ""))
+                        arrayOf(
+                            LineData(0, "").apply { touch() },
+                            LineData(1, "").apply {
+                                setJumpsAndSwitches(
+                                    JumpsAndSwitches().apply {
+                                        addSwitch(0, IntArray(1) { 0 })
+                                        fillArrays()
+                                    }
+                                )
+                            },
+                        )
                     )
                 }
             }
@@ -48,12 +70,41 @@ class IntellijRawCoverageDataProviderTest {
         assertSoftly(actualData) {
             size shouldBe 1
             first() shouldBeEqualToComparingFields RawCoverageData {
-                group = sourceName
                 aClass = className
-                branchesTotal = 0
-                branchesCovered = 0
-                linesCovered = 0
-                linesTotal = 1
+                branches(0, 1)
+                lines(1, 2)
+                instr(instrCovered, instrCovered + instrUncovered)
+            }
+        }
+    }
+
+    @Test
+    fun `should return list with single class and no coverage`() {
+        // GIVEN
+        val className = "com.example.Class2"
+        val sourceName = "Class2.java"
+        val provider = IntellijRawCoverageDataProvider(
+            ProjectData().apply {
+                instructions[className] = ClassInstructions(emptyArray())
+
+                getOrCreateClassData(className).apply {
+                    source = sourceName
+                    setLines(arrayOf(LineData(1, "")))
+                }
+            }
+        )
+
+        // WHEN
+        val actualData: List<RawCoverageData> = provider.obtainData()
+
+        // THEN
+        assertSoftly(actualData) {
+            size shouldBe 1
+            first() shouldBeEqualToComparingFields RawCoverageData {
+                aClass = className
+                branches(0, 0)
+                lines(0, 1)
+                instr(0, 0)
             }
         }
     }