+++ delta-coverage-gradle/src/functionalTest/kotlin/io/github/surpsg/deltacoverage/gradle/DeltaCoverageGradleReleasesTest.kt
@@ -33,8 +33,7 @@ class DeltaCoverageGradleReleasesTest {
             "5.6",
             "6.7.1",
             "7.6.4",
-            "8.6",
-            "8.7", // the latest release or release candidate
+            "8.9", // the latest release or release candidate
         ]
     )
     fun `deltaCoverage task should be completed successfully on Gradle release`(