+++ delta-coverage-gradle/src/main/kotlin/io/github/surpsg/deltacoverage/gradle/sources/lookup/KoverPluginSourcesLookup.kt
@@ -1,7 +1,8 @@
 package io.github.surpsg.deltacoverage.gradle.sources.lookup
 
-import io.github.surpsg.deltacoverage.gradle.sources.lookup.SourcesAutoLookup.Companion.newAutoDetectedSources
+import io.github.surpsg.deltacoverage.gradle.utils.lazyFileCollection
 import org.gradle.api.Task
+import org.gradle.api.file.FileCollection
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.nio.file.FileSystem
@@ -16,52 +17,49 @@ internal class KoverPluginSourcesLookup(
     lookupContext: SourcesAutoLookup.Context
 ) : CacheableLookupSources(lookupContext) {
 
-    override fun lookupSources(lookupContext: SourcesAutoLookup.Context): SourcesAutoLookup.AutoDetectedSources {
-        return lookupContext.project.allprojects.asSequence()
-            .map { it.tasks.findByName(KOVER_GENERATE_ARTIFACTS_TASK_NAME) }
-            .filterNotNull()
-            .fold(lookupContext.objectFactory.newAutoDetectedSources()) { koverInputs, koverGenerateArtifactsTask ->
-                log.debug("Found Kover configuration in gradle project '{}'", koverGenerateArtifactsTask.project.name)
+    override fun lookupCoverageBinaries(lookupContext: SourcesAutoLookup.Context): FileCollection {
+        return lookupContext.project.lazyFileCollection {
+            lookupContext.project.allprojects.asSequence()
+                .map { it.tasks.findByName(KOVER_GENERATE_ARTIFACTS_TASK_NAME) }
+                .filterNotNull()
+                .fold(setOf<String>()) { allBinaries, koverGenerateArtifactsTask ->
+                    log.debug(
+                        "Found Kover configuration in gradle project '{}'",
+                        koverGenerateArtifactsTask.project.name
+                    )
 
-                koverInputs.apply {
-                    applyKoverOutputs(koverGenerateArtifactsTask)
+                    allBinaries + obtainKoverBinaries(koverGenerateArtifactsTask)
                 }
-            }
+                .let { lookupContext.project.files(it) }
+        }
     }
 
-    private fun SourcesAutoLookup.AutoDetectedSources.applyKoverOutputs(
+    private fun obtainKoverBinaries(
         koverGenerateArtifactsTask: Task
-    ) {
+    ): Set<String> {
         val rootProjectPath: Path = koverGenerateArtifactsTask.resolveRootProjectDirPath()
-        koverGenerateArtifactsTask.outputs
+        return koverGenerateArtifactsTask.outputs
             .files
             .asSequence()
             .map { fileSystem.getPath(it.absolutePath) }
             .filter { it.name.endsWith(KOVER_ARTIFACTS_FILE_NAME) }
             .take(1)
             .mapNotNull { it.parseArtifactFile(rootProjectPath) }
-            .forEach {
-                allBinaryCoverageFiles.from(it.coverageBinaries)
-                allClasses.from(it.classFiles)
-                allSources.from(it.sources)
-            }
+            .flatMap { it }
+            .toSet()
     }
 
     private fun Task.resolveRootProjectDirPath(): Path =
         fileSystem.getPath(project.rootProject.layout.projectDirectory.asFile.absolutePath)
 
-    private fun Path.parseArtifactFile(rootProjectPath: Path): KoverArtifacts? = if (exists()) {
+    private fun Path.parseArtifactFile(rootProjectPath: Path): Set<String>? = if (exists()) {
         val iterator: Iterator<String> = readLines().iterator()
 
-        val sources: Set<String> = iterator.readArtifactsSection(rootProjectPath)
-        val outputs: Set<String> = iterator.readArtifactsSection(rootProjectPath)
-        val coverageBinaries: Set<String> = iterator.readArtifactsSection(rootProjectPath)
+        iterator.readArtifactsSection(rootProjectPath)
+        iterator.readArtifactsSection(rootProjectPath)
 
-        KoverArtifacts(
-            coverageBinaries = coverageBinaries,
-            classFiles = outputs,
-            sources = sources,
-        )
+        val coverageBinaries: Set<String> = iterator.readArtifactsSection(rootProjectPath)
+        coverageBinaries
     } else {
         null
     }
@@ -73,12 +71,6 @@ internal class KoverPluginSourcesLookup(
             .map { it.absolutePathString() }
             .toSet()
     }
-
-    private data class KoverArtifacts(
-        val coverageBinaries: Set<String>,
-        val classFiles: Set<String>,
-        val sources: Set<String>,
-    )
 
     private fun Iterator<String>.readUntil(
         interruptCondition: (String) -> Boolean