+++ delta-coverage-gradle/src/test/kotlin/io/github/surpsg/deltacoverage/gradle/sources/lookup/JacocoPluginSourcesLookupTest.kt
@@ -3,25 +3,33 @@ package io.github.surpsg.deltacoverage.gradle.sources.lookup
 import io.github.surpsg.deltacoverage.gradle.DeltaCoverageConfiguration
 import io.github.surpsg.deltacoverage.gradle.sources.SourceType
 import io.github.surpsg.deltacoverage.gradle.unittest.testJavaProject
+import io.kotest.assertions.assertSoftly
 import io.kotest.matchers.collections.shouldHaveAtLeastSize
 import io.kotest.matchers.collections.shouldHaveSize
+import io.kotest.matchers.string.shouldEndWith
 import org.gradle.api.Project
 import org.gradle.api.file.FileCollection
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.EnumSource
+import org.junit.jupiter.params.provider.Arguments
+import org.junit.jupiter.params.provider.Arguments.arguments
+import org.junit.jupiter.params.provider.MethodSource
 
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 internal class JacocoPluginSourcesLookupTest {
 
+    private val project: Project = testJavaProject {
+        pluginManager.apply("jacoco")
+    }
+
     @ParameterizedTest
-    @EnumSource(SourceType::class)
+    @MethodSource("sourcesParameters")
     fun `should return source if source found in jacoco configuration`(
-        sourceType: SourceType
+        sourceType: SourceType,
+        expectedFile: String,
     ) {
         // GIVEN
-        val project: Project = testJavaProject {
-            pluginManager.apply("jacoco")
-        }
-
         val sourcesLookup = JacocoPluginSourcesLookup(
             SourcesAutoLookup.Context(
                 project,
@@ -33,14 +41,14 @@ internal class JacocoPluginSourcesLookupTest {
         val actualSources: FileCollection = sourcesLookup.lookup(sourceType)
 
         // THEN
-        actualSources shouldHaveAtLeastSize 1
+        assertSoftly {
+            actualSources shouldHaveAtLeastSize 1
+            actualSources.first().absolutePath shouldEndWith expectedFile
+        }
     }
 
-    @ParameterizedTest
-    @EnumSource(SourceType::class)
-    fun `should return empty source if jacoco not found in project`(
-        sourceType: SourceType
-    ) {
+    @Test
+    fun `should return empty source if jacoco not found in project`() {
         // GIVEN
         val project: Project = testJavaProject()
 
@@ -52,9 +60,16 @@ internal class JacocoPluginSourcesLookupTest {
             )
         )
         // WHEN
-        val actualSources: FileCollection = sourcesLookup.lookup(sourceType)
+        val actualSources: FileCollection = sourcesLookup.lookup(SourceType.COVERAGE_BINARIES)
 
         // THEN
         actualSources shouldHaveSize 0
     }
+
+    @Suppress("UnusedPrivateMember")
+    private fun sourcesParameters(): List<Arguments> = listOf(
+        arguments(SourceType.COVERAGE_BINARIES, "/test.exec"),
+        arguments(SourceType.CLASSES, "/build/classes/java/main"),
+        arguments(SourceType.SOURCES, "/src/main/java"),
+    )
 }