+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Delta-Coverage Gradle plugin Changelog
 
+## 2.4.0
+
+### Changed
+- Source code paths are inferred from sourceSets if no custom paths are specified.
+- `Markdown` and `Console` reports:
+  - now don't have `source` column.
+  - now contain min coverage data.
+  - (console) contain `✔`/`✖` regards to coverage check success or failure.
+  - (markdown) contain 🟢/🔴 regards to coverage check success or failure.
+- Console report now enabled by the default.
+
+### Added
+- Added GitHub Action for posting delta coverage report to PR comment. See [docs](./actions/delta-coverage-report/README.md)
+
+
 ## 2.3.0
 
 ### Added